Camera: remove fine/normal/superfine jpegquality and use always int
We use now a set of int values the user can exactly assign. Showing as
well the normal fine superfine as choice is just a duplicate. As well
we can get rid of the weird converts of the mixed values.

To take care that the user does not get into problems we updated the db
version and we make a convertion.

Change-Id: I2615666bc1c929c0c045840fad1ece55fb0640af